requirements.txt
Pillow==10.2.0
click==8.1.7